Updated file names so that constructor matches file name additionally also refactored main.js so that is sync with filenames
