+++ src/main.js
@@ -1,8 +1,8 @@
-var scroll = require('./documentWatcher.js');
-var GrabInput = require('./userinput/obtainInput.js');
-var GrabStyles = require('./grabStyles.js');
-var dommutate = require('./domMutations.js');
-var Pointer = require('./pointerWatcher.js');
+var DocumentWatcher = require('./documentWatcher.js');
+var GrabInput = require('./userinput/grabInput.js');
+var GrabExtSheets = require('./GrabExtSheets.js');
+var PointerWatcher = require('./pointerWatcher.js');
+var domMutationsWatcher = require('./domMutations.js');
 var utils = require('./utils.js');
 
 
@@ -10,17 +10,37 @@ var inputScript = new GrabInput();
 console.log("Get all tags:", inputScript.grabAllTags());
 console.log("Add listeners to input tags:", inputScript.addListeners());
 
-//
-// var grabPageStyle = new GrabStyles();
-//
-//
-// console.log("gettags",test2.getTags());
-// console.log("getStyleSheetsLink",test2.getStyleSheetsLink());
-// console.log("getStyleSheetsStyle",test2.getStyleSheetsStyle());
-// console.log("getStyleSheetsStyle",test2.getScripts());
-//
-//
-// var poi = new Pointer();
-// console.log("bound pointer", poi.bindPointer());
+
+var grabSheets = new GrabExtSheets();
+console.log("gettags", grabSheets.getTags());
+console.log("getStyleSheetsLink", grabSheets.getStyleSheetsLink());
+console.log("getStyleSheetsStyle", grabSheets.getStyleSheetsStyle());
+console.log("getScripts", grabSheets.getScripts());
+console.log("Send scripts and sheets to storage", grabSheets.sendToStorage());
+
+
+var pointerTracker = new PointerWatcher();
+console.log("bound pointer", pointerTracker.bindPointer());
 
 dommutate.getInstance();
+
+
+
+
+// var inputScript = new GrabInput();
+// var scriptMethods = [grabAllTags, addListeners];
+//
+// var grabPageStyle = new GrabStyles();
+// var methodList = [getTags, getStyleSheetsLink, getStyleSheetsStyle, getScripts];
+//
+// var runFunctions= function(constrcter, methods){
+//   for(var i = 0; i < methods.length; i++){
+//     var it =  constrcter[methods[i]]
+//     console.log("waht is it", it);
+//   }
+// };
+//
+// runFunctions(new GrabInput(), scriptMethods);
+// runFunctions(new new GrabStyles(), methodList);
+
+
